requirements.txt
streamlit==1.32.0
python-pptx==0.6.23
moviepy==1.0.3
pillow==10.2.0
gTTS==2.5.1
python-docx==1.1.0
numpy==1.26.4
opencv-python==4.9.0.80
pywin32==306
azure-cognitiveservices-speech==1.34.0
openai==1.12.0
requests==2.31.0